[gdf] Rename nose_answer tests
